Consolidate game.core & game.server into customExperiment.js
