+++ experiments/SpatialExperiment/package.json
@@ -21,6 +21,7 @@
     "node-uuid": "1.4.1",
     "request": "^2.87.0",
     "socket.io": "^2.1.1",
+    "socket.io-client": "^2.2.0",
     "underscore": "~1.7.0",
     "webpack": "^4.28.3",
     "xmldom": "^0.1.27"
